Fix warning (un/signed comparison)
git-svn-id: http://root.cern.ch/svn/root/trunk@46500 27541ba8-7e3a-0410-8455-c3a389f83636

+++ lib/Interpreter/StoredValueRef.cpp
@@ -17,7 +17,7 @@ m_Mem(0) {
   if (!(t->isIntegralOrEnumerationType()
         || t->isRealFloatingType()
         || t->hasPointerRepresentation())) {
-    const int64_t size = getAllocSizeInBytes(ctx);
+    const uint64_t size = (uint64_t)getAllocSizeInBytes(ctx);
     if (size > sizeof(m_Buf))
       m_Mem = new char[size];
     else m_Mem = m_Buf;
